+++ cableprobe/analysis.py
@@ -46,6 +46,8 @@ VOLATILE_KEYS = {
     "delta_ma",
     "bus_voltage_v",
     "baseline_ma",
+    "local",
+    "state",
 }
 
 _ADD_ACTIONS = {"add", "bind", "online"}